:bug: Fixed reviewOppRecommendation :hammer: Moved helper functions for the Newsletter into the separate file admin-analytics-newsletter-helper-functions.ts :pencil2: Improved typings :hammer: Refactored AdminAnalyticsNewsletterMessagePreviewWidget.tsx into a pure function component from class component
