+++ app/imports/api/analytic/Email.methods.ts
@@ -17,8 +17,7 @@ export const sendEmailMethod = new ValidatedMethod({
     this.unblock();
     if (!this.userId) {
       throw new Meteor.Error('unauthorized', 'You must be logged in to send emails.', Error().stack);
-    } else
-    if (!Roles.userIsInRole(this.userId, [ROLE.ADMIN])) {
+    } else if (!Roles.userIsInRole(this.userId, [ROLE.ADMIN])) {
       throw new Meteor.Error('unauthorized', 'You must be an Admin to send emails.', Error().stack);
     }
     sendEmail(emailData);
