+++ app/private/database/snapshot/2020-01-24-11-37-22.json
@@ -13867,7 +13867,7 @@
           "careerGoals": [
             "software-developer"
           ],
-          "level": 6,
+          "level": 5,
           "favoriteAcademicPlans": [],
           "favoriteCourses": [],
           "favoriteOpportunities": [],
@@ -13991,7 +13991,7 @@
             "startup-cofounder",
             "ux-designer"
           ],
-          "level": 6,
+          "level": 5,
           "favoriteAcademicPlans": [],
           "favoriteCourses": [],
           "favoriteOpportunities": [],
@@ -14046,7 +14046,7 @@
           "careerGoals": [
             "full-stack-developer"
           ],
-          "level": 6,
+          "level": 5,
           "favoriteAcademicPlans": [],
           "favoriteCourses": [],
           "favoriteOpportunities": [],
@@ -14103,7 +14103,7 @@
           "careerGoals": [
             "data-scientist"
           ],
-          "level": 3,
+          "level": 4,
           "favoriteAcademicPlans": [],
           "favoriteCourses": [],
           "favoriteOpportunities": [],
